% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentix.R
\docType{data}
\name{sentix_file}
\alias{sentix_file}
\title{Sentix}
\format{An object of class \code{character} of length 1.}
\usage{
sentix_file
}
\description{
A package to compute the sentiment polarity of Italian text using the Sentix lexicon.
}
\keyword{sentiment}
